Implemented rudimentary text input system, for use with specifying new scene lengths.

+++ inc/types.hpp
@@ -4,7 +4,7 @@
 	
 	namespace uInt32{
 		void write(FILE *out,uint32_t v);
-		uint16_t read(FILE *in);
+		uint32_t read(FILE *in);
 	}
 	
 	namespace uInt16{
